[✨feat]: MainPage의 테스트용 컴포넌트를 TestPage로 분리
